+++ test/test_outputLayer.py
@@ -16,9 +16,9 @@ class TestOutputLayer(TestCase):
     def test_call_verifyOutputShape(self):
         # Arrange
         layer = OutputLayer(l2_rate=3e-7)
-        m0 = np.random.rand(self.BATCH_SIZE, self.N_CONTEXT, self.N_DIM).astype(np.float32)
-        m1 = np.random.rand(self.BATCH_SIZE, self.N_CONTEXT, self.N_DIM).astype(np.float32)
-        m2 = np.random.rand(self.BATCH_SIZE, self.N_CONTEXT, self.N_DIM).astype(np.float32)
+        m0 = np.random.rand(self.BATCH_SIZE, self.N_CONTEXT, self.N_DIM).astype(np.float16)
+        m1 = np.random.rand(self.BATCH_SIZE, self.N_CONTEXT, self.N_DIM).astype(np.float16)
+        m2 = np.random.rand(self.BATCH_SIZE, self.N_CONTEXT, self.N_DIM).astype(np.float16)
 
         # Act
         result = layer([m0, m1, m2])
